first try at refactoring handlers into golang philosophy style
